+++ Source/General/ApplicationSettings.vb
@@ -406,10 +406,12 @@ Custom... = $enter_text:Enter a custom Pixel Aspect Ratio.$"
 
 3840 x auto = 3840 x 0
 3200 x auto = 3200 x 0
+2880 x auto = 2880 x 0
 2560 x auto = 2560 x 0
 
 1920 x auto = 1920 x 0
-1680 x auto = 1680 x 0
+1600 x auto = 1600 x 0
+1440 x auto = 1440 x 0
 1280 x auto = 1280 x 0
 960 x auto = 960 x 0